+++ charts/lib/src/core/a11y/a11y_node.dart
@@ -22,7 +22,7 @@ class A11yNode {
   A11yNode(this.label, this.boundingBox, {this.onFocus});
 
   /// The bounding box for this node.
-  final Rectangle<int> boundingBox;
+  final Rectangle<double> boundingBox;
 
   /// The textual description of this node.
   final String label;
